+++ src/main/java/com/kinoticket/backend/dto/CinemaHallDTO.java
@@ -1,4 +1,6 @@
-package com.kinoticket.backend.model;
+package com.kinoticket.backend.dto;
+
+import com.kinoticket.backend.model.CinemaHall;
 
 import lombok.Data;
