+++ src/main/java/com/kinoticket/backend/dto/FilmShowInformationDTO.java
@@ -1,8 +1,11 @@
-package com.kinoticket.backend.model;
+package com.kinoticket.backend.dto;
 
 import java.sql.Time;
 import java.util.Date;
 import java.util.List;
+
+import com.kinoticket.backend.model.FilmShow;
+import com.kinoticket.backend.model.FilmShowSeat;
 
 import lombok.Data;
