+++ src/main/java/com/kinoticket/backend/rest/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.kinoticket.backend.dto;
+package com.kinoticket.backend.rest.request;
 
 import java.util.Set;
