+++ src/main/java/com/kinoticket/backend/rest/response/MessageResponse.java
@@ -1,4 +1,4 @@
-package com.kinoticket.backend.dto;
+package com.kinoticket.backend.rest.response;
 
 public class MessageResponse {
   private String message;
