+++ src/main/java/com/kinoticket/backend/service/BookingService.java
@@ -7,8 +7,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import com.kinoticket.backend.Exceptions.EntityNotFound;
 import com.kinoticket.backend.Exceptions.MissingParameterException;
+import com.kinoticket.backend.dto.BookingDTO;
 import com.kinoticket.backend.model.Booking;
-import com.kinoticket.backend.model.BookingDTO;
 import com.kinoticket.backend.model.FilmShowSeat;
 import com.kinoticket.backend.model.Ticket;
 import com.kinoticket.backend.repositories.BookingRepository;
